# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AssetMind Backend API",
        "description": "REST API untuk Sistem Pendukung Keputusan Perawatan Aset Industri Perkebunan Kelapa Sawit. Backend Laravel terintegrasi dengan AI Engine (FastAPI) untuk embedding, chatbot RAG, dan rekomendasi pemeliharaan.",
        "contact": {
            "name": "AssetMind",
            "email": "firasyan.daffa123@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server lokal (Laravel artisan serve)"
        }
    ],
    "paths": {
...
        "/api/ingestion/repair-reports/{id}": {
            "post": {
                "tags": [
                    "Knowledge"
                ],
                "summary": "Ingest catatan perbaikan ke knowledge base (khusus teknik)",
                "operationId": "ingestRepairReport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Catatan perbaikan berhasil di-ingest"
                    },
                    "500": {
                        "description": "Gagal ingest catatan perbaikan"
                    }
                },
                "security": [
                    {
                        "bearerSanctum": []
                    }
                ]
            }
        },
...
        "/api/repair-reports": {
            "get": {
                "tags": [
                    "Repair Report"
                ],
                "summary": "Daftar catatan perbaikan",
                "description": "Mekanik hanya melihat catatan miliknya sendiri.",
                "operationId": "repairReportIndex",
                "responses": {
                    "200": {
                        "description": "Data catatan perbaikan berhasil diambil"
                    }
                },
                "security": [
                    {
                        "bearerSanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Repair Report"
                ],
                "summary": "Buat catatan perbaikan (khusus mekanik)",
                "description": "Pengajuan harus milik mekanik & berstatus approved, belum ada catatan, jam jalan sesuai aset.",
                "operationId": "repairReportStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "repair_request_id",
                                    "operating_hour_id",
                                    "tanggal_perbaikan",
                                    "catatan_temuan",
                                    "cara_perbaikan"
                                ],
                                "properties": {
                                    "repair_request_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "operating_hour_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "tanggal_perbaikan": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-10"
                                    },
                                    "catatan_temuan": {
                                        "type": "string",
                                        "example": "Bearing aus"
                                    },
                                    "cara_perbaikan": {
                                        "type": "string",
                                        "example": "Ganti bearing baru"
                                    },
                                    "bukti_foto": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "llm_suggestion": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Catatan perbaikan berhasil dibuat"
                    },
                    "403": {
                        "description": "Bukan milik Anda",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Aturan bisnis dilanggar",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerSanctum": []
                    }
                ]
            }
        },
        "/api/repair-reports/{id}": {
            "get": {
                "tags": [
                    "Repair Report"
                ],
                "summary": "Detail catatan perbaikan",
                "operationId": "repairReportShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail catatan perbaikan berhasil diambil"
                    },
                    "403": {
                        "description": "Bukan milik Anda",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerSanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Repair Report"
                ],
                "summary": "Perbarui catatan perbaikan",
                "operationId": "repairReportUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "operating_hour_id",
                                    "tanggal_perbaikan",
                                    "catatan_temuan",
                                    "cara_perbaikan"
                                ],
                                "properties": {
                                    "operating_hour_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "tanggal_perbaikan": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-10"
                                    },
                                    "catatan_temuan": {
                                        "type": "string"
                                    },
                                    "cara_perbaikan": {
                                        "type": "string"
                                    },
                                    "bukti_foto": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "llm_suggestion": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Catatan perbaikan berhasil diperbarui"
                    },
                    "403": {
                        "description": "Bukan milik Anda",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Aturan bisnis dilanggar",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerSanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Repair Report"
                ],
                "summary": "Hapus catatan perbaikan",
                "operationId": "repairReportDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Catatan perbaikan berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Bukan milik Anda",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerSanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "MessageResponse": {
                "properties": {
                    "message": {
                        "description": "Definisi dasar dokumentasi OpenAPI/Swagger untuk AssetMind Backend API.\n\nMemakai PHP 8 Attributes (didukung swagger-php v6).\nAnotasi path tersebar di controller masing-masing.",
                        "type": "string",
                        "example": "Operasi berhasil"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ForbiddenError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Forbidden. Data ini bukan milik Anda."
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerSanctum": {
                "type": "http",
                "description": "Masukkan token Sanctum dari endpoint /auth/login (tanpa kata 'Bearer').",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}